Update: move `CellMask` to types
